backend/requirements.txt
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
PyJWT>=2.8.0,<3.0.0
supabase>=2.3.0,<3.0.0
postgrest>=0.16.0,<1.0.0
httpx>=0.26.0,<1.0.0
python-multipart>=0.0.6,<1.0.0
python-dotenv>=1.0.0,<2.0.0
